requirements.txt
streamlit
pandas
streamlit-extras
mysql-connector-python
python-decouple
plotly